Extract shared mutation onError and user-reload helpers
The react-query mutation hooks repeated two blocks nearly verbatim. Extract
each into a small shared helper:

- getErrorAlert(title, error): builds the standard error-type IAppAlert for a
  failed mutation. Applied to the 6 hooks that showed an error alert
  (add/edit journal, modify permissions, modify schedule, move entry, upsert
  entry). This also normalizes error-message extraction, which had drifted
  between error.message, error.toString() and a mixed fallback, onto a single
  rule (Error.message, else String(error)). For our ApiError that yields
  "<status>: <message>", dropping the redundant "Error: " prefix that
  toString() added. It also fixes two handlers that mistyped the thrown error
  as IAppAlert.

- useReloadUser(): encapsulates invalidate modifyUser -> getCurrentUser ->
  setUser, used by the favorites, tag-update and edit-journal (tag) mutations.
  The two favorites hooks had byte-for-byte identical bodies; both now reduce
  to a mutationFn plus onSuccess: () => reloadUser().

No user-facing behavior change beyond the normalized error message text.
Full unit suite green (18 files / 116 tests); knip reports no unused code.
